+++ src/use/helpers/Toasts.ts
@@ -1,8 +1,10 @@
-//import ReportBug from "@/components/common/ReportBug.vue";
-//import { markRaw } from "vue";
 import { useToast } from "@/toast";
 
 const toast = useToast();
+
+export function showOperationCreatedToast(operationType: String) {
+    toast.info(`Successfully created ${operationType} request. You can check its current state in your dashboard on the welcome page.`);
+}
 
 export function showNetworkErrorToast() {
     toast.error("Network error: make sure to establish a stable internet connection.");
@@ -12,7 +14,6 @@ export function showAPIToast(code: number, msg?: string) {
     switch (code) {
         case 400:
         case 500:
-            //toast.error(markRaw(ReportBug));
             toast.error("Something went wrong on our side. Please consider reporting this bug.");
             break;
         case 403: